+++ game/settings/settings.py
@@ -959,6 +959,7 @@ class Settings:
     enable_base_capture_cheat: bool = False
     enable_transfer_cheat: bool = False
     enable_runway_state_cheat: bool = False
+    enable_air_wing_adjustments: bool = False
 
     # LUA Plugins system
     plugins: Dict[str, bool] = field(default_factory=dict)