+++ qt_ui/windows/AirWingConfigurationDialog.py
@@ -12,7 +12,6 @@ from PySide6.QtGui import QIcon, QStandardItem, QStandardItemModel
 from PySide6.QtWidgets import (
     QComboBox,
     QDialog,
-    QGroupBox,
     QHBoxLayout,
     QLabel,
     QLineEdit,
@@ -29,6 +28,7 @@ from PySide6.QtWidgets import (
     QToolButton,
     QMessageBox,
     QSpinBox,
+    QGroupBox,
 )
 
 from game import Game
@@ -39,7 +39,6 @@ from game.dcs.aircrafttype import AircraftType
 from game.squadrons import AirWing, Pilot, Squadron
 from game.squadrons.squadrondef import SquadronDef
 from game.theater import ControlPoint, ParkingType
-from game.theater.start_generator import GeneratorSettings
 from qt_ui.uiconstants import AIRCRAFT_ICONS, ICONS
 from qt_ui.widgets.combos.QSquadronLiverySelector import SquadronLiverySelector
 from qt_ui.widgets.combos.primarytaskselector import PrimaryTaskSelector
@@ -196,14 +195,14 @@ class SquadronConfigurationBox(QGroupBox):
         coalition: Coalition,
         squadron: Squadron,
         parking_tracker: AirWingConfigParkingTracker,
-        gen_settings: GeneratorSettings,
+        aircraft_present: bool,
     ) -> None:
         super().__init__()
         self.game = game
         self.coalition = coalition
         self.squadron = squadron
         self.parking_tracker = parking_tracker
-        self.gen_settings = gen_settings
+        self.aircraft_present = aircraft_present
 
         columns = QHBoxLayout()
         self.setLayout(columns)
@@ -326,7 +325,7 @@ class SquadronConfigurationBox(QGroupBox):
         required_slots = self.parking_tracker.used_parking_at(self.squadron.location)
         required_slots_string = (
             f"Parking slots required: {required_slots}<br/>"
-            if self.gen_settings.squadrons_start_full
+            if self.aircraft_present
             else ""
         )
         total_slots = self.squadron.location.total_aircraft_parking(
@@ -446,14 +445,14 @@ class SquadronConfigurationLayout(QVBoxLayout):
         coalition: Coalition,
         squadrons: list[Squadron],
         parking_tracker: AirWingConfigParkingTracker,
-        gen_settings: GeneratorSettings,
+        aircraft_present: bool,
     ) -> None:
         super().__init__()
         self.game = game
         self.coalition = coalition
         self.squadron_configs = []
         self.parking_tracker = parking_tracker
-        self.gen_settings = gen_settings
+        self.aircraft_present = aircraft_present
         for squadron in squadrons:
             self.add_squadron(squadron)
 
@@ -476,7 +475,11 @@ class SquadronConfigurationLayout(QVBoxLayout):
 
     def add_squadron(self, squadron: Squadron) -> None:
         squadron_config = SquadronConfigurationBox(
-            self.game, self.coalition, squadron, self.parking_tracker, self.gen_settings
+            self.game,
+            self.coalition,
+            squadron,
+            self.parking_tracker,
+            self.aircraft_present,
         )
         squadron_config.remove_squadron_signal.connect(self.remove_squadron)
         self.squadron_configs.append(squadron_config)
@@ -493,14 +496,14 @@ class AircraftSquadronsPage(QWidget):
         coalition: Coalition,
         squadrons: list[Squadron],
         parking_tracker: AirWingConfigParkingTracker,
-        gen_settings: GeneratorSettings,
+        aircraft_present: bool,
     ) -> None:
         super().__init__()
         layout = QVBoxLayout()
         self.setLayout(layout)
 
         self.squadrons_config = SquadronConfigurationLayout(
-            game, coalition, squadrons, parking_tracker, gen_settings
+            game, coalition, squadrons, parking_tracker, aircraft_present
         )
         self.squadrons_config.config_changed.connect(self.on_squadron_config_changed)
 
@@ -536,14 +539,14 @@ class AircraftSquadronsPanel(QStackedLayout):
         game: Game,
         coalition: Coalition,
         parking_tracker: AirWingConfigParkingTracker,
-        gen_settings: GeneratorSettings,
+        aircraft_present: bool,
     ) -> None:
         super().__init__()
         self.game = game
         self.coalition = coalition
         self.parking_tracker = parking_tracker
         self.squadrons_pages: dict[AircraftType, AircraftSquadronsPage] = {}
-        self.gen_settings = gen_settings
+        self.aircraft_present = aircraft_present
         for aircraft, squadrons in self.air_wing.squadrons.items():
             self.new_page_for_type(aircraft, squadrons)
 
@@ -567,7 +570,7 @@ class AircraftSquadronsPanel(QStackedLayout):
             self.coalition,
             squadrons,
             self.parking_tracker,
-            self.gen_settings,
+            self.aircraft_present,
         )
         page.remove_squadron_page.connect(self.remove_page_for_type)
         self.addWidget(page)
@@ -657,7 +660,7 @@ class AircraftTypeList(QListView):
 
 class AirWingConfigurationTab(QWidget):
     def __init__(
-        self, coalition: Coalition, game: Game, gen_settings: GeneratorSettings
+        self, coalition: Coalition, game: Game, aircraft_present: bool
     ) -> None:
         super().__init__()
 
@@ -678,7 +681,7 @@ class AirWingConfigurationTab(QWidget):
         layout.addWidget(add_button, 2, 1, 1, 1)
 
         self.squadrons_panel = AircraftSquadronsPanel(
-            game, coalition, self.parking_tracker, gen_settings
+            game, coalition, self.parking_tracker, aircraft_present
         )
         self.squadrons_panel.page_removed.connect(self.type_list.remove_aircraft_type)
         layout.addLayout(self.squadrons_panel, 1, 3, 2, 1)
@@ -752,7 +755,7 @@ class AirWingConfigurationTab(QWidget):
 class AirWingConfigurationDialog(QDialog):
     """Dialog window for air wing configuration."""
 
-    def __init__(self, game: Game, gen_setting: GeneratorSettings, parent) -> None:
+    def __init__(self, game: Game, aircraft_present: bool, parent) -> None:
         super().__init__(parent)
         self.setMinimumSize(1024, 768)
         self.setWindowTitle(f"Air Wing Configuration")
@@ -763,7 +766,8 @@ class AirWingConfigurationDialog(QDialog):
 
         doc_label = QLabel(
             "Use this opportunity to customize the squadrons available to your "
-            "coalition. <strong>This is your only opportunity to make changes.</strong>"
+            "coalition. You can make changes at a later stage if "
+            'the "Enable Air Wing adjustments" cheat option is enabled.'
             "<br /><br />"
             "To accept your changes and continue, close this window."
         )
@@ -775,7 +779,7 @@ class AirWingConfigurationDialog(QDialog):
 
         self.tabs = []
         for coalition in game.coalitions:
-            coalition_tab = AirWingConfigurationTab(coalition, game, gen_setting)
+            coalition_tab = AirWingConfigurationTab(coalition, game, aircraft_present)
             name = "Blue" if coalition.player else "Red"
             self.tab_widget.addTab(coalition_tab, name)
             self.tabs.append(coalition_tab)
@@ -804,7 +808,7 @@ class AirWingConfigurationDialog(QDialog):
         result = QMessageBox.information(
             None,
             "Discard changes?",
-            "Are you sure you want to discard your changes and start the campaign?",
+            "Are you sure you want to discard your changes?",
             QMessageBox.StandardButton.Yes,
             QMessageBox.StandardButton.No,
         )